requirements.txt
scipy
numpy
pycryptodome
reedsolo
pywavelets